chore: organize namespaces
