+++ src/Input/InputInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Symblaze\Console\IO;
+namespace Symblaze\Console\Input;
 
 interface InputInterface
 {
