+++ src/Input/InputTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Symblaze\Console\IO;
+namespace Symblaze\Console\Input;
 
 /**
  * A collection of methods to interact with the input.
